refactor: streamline artifact information handling in GitHub workflow
This commit refactors the way artifact information is handled in the release_and_publish.yml GitHub workflow. The get_os function has been moved inside the Set matrix step for better encapsulation. Artifact information is now stored as a string instead of an associative array, and jq is used to generate JSON directly from shell variables. This simplifies the code and improves readability.