feat: automate package.json version update in GitHub workflow
This commit replaces the manual script for updating the package.json version with a new GitHub action. This action is defined in 'update_package_json_version.yml' and is now used within the release_and_publish workflow. The change simplifies the process of keeping the package.json version synchronized with cargo's version.