feat: Update paseto_cli version to 1.0.0-alpha.13 and improve binary download error handling
This commit updates the `paseto_cli` version in both Cargo.lock and Cargo.toml files from 1.0.0-alpha.11 to 1.0.0-alpha.13.

In addition, it modifies the default BINARY_DISTRIBUTION_VERSION in install.js from '1.0.0-alpha.2' to '1.0.0-alpha.11'.

The binary download process is also improved by adding a try-catch block for better error handling during the download and installation of binaries.

Lastly, minor changes are made to console log messages for clarity when a platform-specific package is already installed.