feat: Update paseto_cli version and improve binary installation script
This commit updates the `paseto_cli` package from version `1.0.0-alpha.18` to `1.0.0-alpha.19`. It also refactors the binary installation script in `install.js`, improving error handling, adding comments for clarity, and making minor adjustments such as changing single quotes to double quotes for consistency.

In addition, it changes the path of the executable in the "bin" field of `package.json` from "bin/paseto_cli" to "bin/cli".